docs: podman-top: fix nonworking example
(minor) One example in the podman-top man page results in an
error when actually invoked:

    unable to find PID field in ps output. try a different set of ps arguments

Cause: commit 6cb1c31d (PR #400) has to check PIDs in order to
filter those in the container.

Solution: add 'pid' to the list of requested output fields in
the sample command.

Signed-off-by: Ed Santiago <santiago@redhat.com>

Closes: #457
Approved by: rhatdan