Copy core buildah content into libpod/buildah
Eventually libpod/podman will want to use both commit and build from buildah's code.  And then
in turn, buildah would vendor in libpod for the core content as well.  This will lead to tighter
integration between the projects and less code duplication.

Once this PR is merged, I will vendor in this code to buildah.

Signed-off-by: baude <bbaude@redhat.com>