vendor.conf: Bump CNI to v0.6.0
Neither the nominal version nor vendored content had changed since
a031b83a (Initial checkin from CRI-O repo, 2017-11-01):

  $ git ls-tree origin/master -- vendor/github.com/containernetworking/cni
  040000 tree 744c091cc1dbb869b2cf714cf6b5e8d33fe17d9c    vendor/github.com/containernetworking/cni
  $ git ls-tree a031b83a -- vendor/github.com/containernetworking/cni
  040000 tree 744c091cc1dbb869b2cf714cf6b5e8d33fe17d9c    vendor/github.com/containernetworking/cni
  $ git grep /cni a031b83a -- vendor.conf
  a031b83a:vendor.conf:github.com/containernetworking/cni v0.4.0

I'm not quite sure which upstream version the old vendored content
came from, but it certainly wasn't v0.4.0.  I've bumped our nominal
version to v0.6.0 and re-vendored with:

  $ vndr github.com/containernetworking/cni
  $ git add -A vendor/github.com/containernetworking/cni

to generate this commit.  The only change to the vendored content is
text in the README (unlikely to break anything ;), and sitting on an
upstream tag is nice, so I think this is good enough.

Signed-off-by: W. Trevor King <wking@tremily.us>